+++ crypto-onramp-example/src/main/java/com/stripe/android/crypto/onramp/example/OnrampConfigurationFactory.kt
@@ -44,10 +44,16 @@ internal object OnrampConfigurationFactory {
                     existingPaymentMethodRequired = false
                 )
             )
+            .samsungPayConfig(
+                OnrampConfiguration.SamsungPayConfig(
+                    serviceId = SAMSUNG_PAY_SERVICE_ID,
+                )
+            )
     }
 }
 
 private const val MERCHANT_NAME = "Onramp Example"
+private const val SAMSUNG_PAY_SERVICE_ID = "SERVICE_ID" // Replace with valid service ID for Samsung Pay.
 
 @Suppress("MaxLineLength")
 private const val PUBLISHABLE_KEY =